qt6 port changes, used to build against qt6.0.0 on windows

+++ Code/Framework/AzQtComponents/AzQtComponents/Components/Widgets/Internal/OverlayWidgetLayer.cpp
@@ -211,7 +211,7 @@ namespace AzQtComponents
 
         void OverlayWidgetLayer::ButtonClicked(size_t index)
         {
-            AZ_Assert(index < m_buttons.size(), "Invalid index for button in overlay layer.");
+            AZ_Assert(index < static_cast<size_t>(m_buttons.size()), "Invalid index for button in overlay layer.");
 
             Button& button = m_buttons[static_cast<int>(index)];
             if (button.m_enabledCheck)
